update test layout
