+++ .gitignore
@@ -1,16 +1,22 @@
-**/.vs
+.idea
+.code
+**/vendor/
 .pulumi
 **/bin/
 **/obj/
+Pulumi.*.yaml
 **/node_modules/
-**/vendor/
-**/Pulumi.*.yaml
-**/yarn-error.log
-yarn.lock
-venv/
+
+**/command-output/
+
 .idea/
 *.iml
-*.egg-info
-__pycache__
-.mypy_cache
+
+yarn.lock
 **/pulumiManifest.go
+
+ci-scripts
+**/schema.go
+**/version.txt
+**/nuget
+**/dist
