--- .travis.yml
@@ -1,32 +1,0 @@
-# It may be tempting to add parens around each individual clause in this expression, but Travis then builds pushes anyway
-if: branch = master OR branch =~ ^features/ OR branch =~ ^feature- OR branch =~ ^release/ OR tag =~ ^v\d+.
-language: go
-go: 1.14.x
-sudo: true # give us 7.5GB and >2 bursted cores.
-git:
-    depth: false
-before_install:
-    - git clone https://github.com/pulumi/scripts ${GOPATH}/src/github.com/pulumi/scripts
-    - source ${GOPATH}/src/github.com/pulumi/scripts/ci/prepare-environment.sh
-    - source ${PULUMI_SCRIPTS}/ci/keep-failed-tests.sh
-    # Install latest docker-ce
-    - curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
-    - sudo add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable"
-    - sudo apt-get update
-    - sudo apt-get -y -o Dpkg::Options::="--force-confnew" install docker-ce
-install:
-    - source ${PULUMI_SCRIPTS}/ci/install-common-toolchain.sh
-    - source ${PULUMI_SCRIPTS}/ci/configure-aws.sh
-    # Use the pulumi-ci AWS profile configured by the above script.
-    # This repo runs tests that require access to create resources in AWS.
-    - export AWS_PROFILE=pulumi-ci
-    - curl -L https://get.pulumi.com/ | bash
-    - export PATH=$HOME/.pulumi/bin:$PATH
-before_script:
-    - ${PULUMI_SCRIPTS}/ci/ensure-dependencies
-script:
-    - make travis_${TRAVIS_EVENT_TYPE}
-after_failure:
-    - ${PULUMI_SCRIPTS}/ci/upload-failed-tests
-notifications:
-    webhooks: https://zlmgkhmhjc.execute-api.us-west-2.amazonaws.com/stage/travis
