--- examples/aws-py/.gitignore
@@ -1,1 +1,0 @@
-*.pyc
