+++ examples/examples_test.go
@@ -15,185 +15,10 @@
 package examples
 
 import (
-	"fmt"
-	"os"
-	"path"
-	"testing"
-
 	"github.com/pulumi/pulumi/pkg/v2/testing/integration"
-
-	"github.com/stretchr/testify/assert"
 )
 
 var base = integration.ProgramTestOptions{
 	ExpectRefreshChanges: true, // Docker resources generally see changes when refreshed.
 	// Note: no Config! This package should be usable without any config.
-}
-
-func TestAws(t *testing.T) {
-	region := os.Getenv("AWS_REGION")
-	if region == "" {
-		t.Skipf("Skipping test due to missing AWS_REGION environment variable")
-	}
-	fmt.Printf("AWS Region: %v\n", region)
-
-	cwd, err := os.Getwd()
-	if !assert.NoError(t, err) {
-		t.FailNow()
-	}
-
-	opts := base.With(integration.ProgramTestOptions{
-		Config: map[string]string{
-			"aws:region": region,
-		},
-		Dependencies: []string{
-			"@pulumi/docker",
-		},
-		Dir: path.Join(cwd, "aws"),
-	})
-	integration.ProgramTest(t, &opts)
-}
-
-func TestAwsPy(t *testing.T) {
-	region := os.Getenv("AWS_REGION")
-	if region == "" {
-		t.Skipf("Skipping test due to missing AWS_REGION environment variable")
-	}
-	fmt.Printf("AWS Region: %v\n", region)
-
-	cwd, err := os.Getwd()
-	if !assert.NoError(t, err) {
-		t.FailNow()
-	}
-
-	opts := base.With(integration.ProgramTestOptions{
-		Config: map[string]string{
-			"aws:region": region,
-		},
-		Dependencies: []string{
-			path.Join("..", "sdk", "python", "bin"),
-		},
-		Dir: path.Join(cwd, "aws-py"),
-	})
-	integration.ProgramTest(t, &opts)
-}
-
-func TestNginx(t *testing.T) {
-	cwd, err := os.Getwd()
-	if !assert.NoError(t, err) {
-		t.FailNow()
-	}
-
-	opts := base.With(integration.ProgramTestOptions{
-		Dependencies: []string{
-			"@pulumi/docker",
-		},
-		Dir: path.Join(cwd, "nginx"),
-	})
-	integration.ProgramTest(t, &opts)
-}
-
-func TestNginxPy(t *testing.T) {
-	cwd, err := os.Getwd()
-	if !assert.NoError(t, err) {
-		t.FailNow()
-	}
-
-	opts := base.With(integration.ProgramTestOptions{
-		Dependencies: []string{
-			path.Join("..", "sdk", "python", "bin"),
-		},
-		Dir: path.Join(cwd, "nginx-py"),
-	})
-	integration.ProgramTest(t, &opts)
-}
-
-func TestNginxCs(t *testing.T) {
-	cwd, err := os.Getwd()
-	if !assert.NoError(t, err) {
-		t.FailNow()
-	}
-
-	opts := base.With(integration.ProgramTestOptions{
-		Dependencies: []string{
-			"Pulumi.Docker",
-		},
-		Dir: path.Join(cwd, "nginx-cs"),
-	})
-	integration.ProgramTest(t, &opts)
-}
-
-func TestNginxGo(t *testing.T) {
-	cwd, err := os.Getwd()
-	if !assert.NoError(t, err) {
-		t.FailNow()
-	}
-
-	opts := base.With(integration.ProgramTestOptions{
-		Dependencies: []string{
-			"github.com/pulumi/pulumi-docker/sdk/v2",
-		},
-		Dir: path.Join(cwd, "nginx-go"),
-	})
-	integration.ProgramTest(t, &opts)
-}
-
-func TestDotNet(t *testing.T) {
-	cwd, err := os.Getwd()
-	if !assert.NoError(t, err) {
-		t.FailNow()
-	}
-
-	opts := base.With(integration.ProgramTestOptions{
-		Dependencies: []string{
-			"Pulumi.Docker",
-		},
-		Dir: path.Join(cwd, "dotnet"),
-	})
-	integration.ProgramTest(t, &opts)
-}
-
-func TestDockerfileWithMultipleTargets(t *testing.T) {
-	cwd, err := os.Getwd()
-	if !assert.NoError(t, err) {
-		t.FailNow()
-	}
-
-	opts := base.With(integration.ProgramTestOptions{
-		Dependencies: []string{
-			"@pulumi/docker",
-		},
-		Dir: path.Join(cwd, "dockerfile-with-targets"),
-	})
-	integration.ProgramTest(t, &opts)
-}
-
-func TestDockerfilePy(t *testing.T) {
-	cwd, err := os.Getwd()
-	if !assert.NoError(t, err) {
-		t.FailNow()
-	}
-
-	opts := base.With(integration.ProgramTestOptions{
-		Dependencies: []string{
-			path.Join("..", "sdk", "python", "bin"),
-		},
-		Dir: path.Join(cwd, "dockerfile-py"),
-	})
-	integration.ProgramTest(t, &opts)
-}
-
-func TestDockerfileGo(t *testing.T) {
-	cwd, err := os.Getwd()
-	if !assert.NoError(t, err) {
-		t.FailNow()
-	}
-
-	opts := base.With(integration.ProgramTestOptions{
-		Dependencies: []string{
-			"github.com/pulumi/pulumi-docker/sdk/v2",
-		},
-		Dir: path.Join(cwd, "dockerfile-go"),
-	})
-	integration.ProgramTest(t, &opts)
 }
