+++ examples/go.mod
@@ -5,6 +5,7 @@ go 1.14
 require (
 	github.com/onsi/ginkgo v1.12.0 // indirect
 	github.com/onsi/gomega v1.9.0 // indirect
+	github.com/pulumi/pulumi/pkg v1.14.1 // indirect
 	github.com/pulumi/pulumi/pkg/v2 v2.0.0
 	github.com/stretchr/testify v1.5.1
 	gopkg.in/airbrake/gobrake.v2 v2.0.9 // indirect
