--- examples/nginx/.gitignore
@@ -1,2 +1,0 @@
-/bin/
-/node_modules/
