--- provider/cmd/pulumi-resource-docker/.gitignore
@@ -1,1 +1,0 @@
-schema.go
