--- sdk/dotnet/.gitignore
@@ -1,4 +1,0 @@
-version.txt
-.vs
-bin
-obj
